Correction of Angle Modulation precision member.
The precision member was allowed to accept values that were less than 0.
This is obviously not correct. Adjusted the class and the test case to
check for this condition.

Algorithm updates related to the change in TopologyVisitors.

+++ src/main/java/net/sourceforge/cilib/measurement/single/diversity/normalisation/TopologyBasedNormalisation.java
@@ -49,7 +49,7 @@ public class TopologyBasedNormalisation extends NormalisationParameter {
 		PopulationBasedAlgorithm algorithm = (PopulationBasedAlgorithm) Algorithm.get();
 		visitor.setDistanceMeasure(distanceMeasure);
 		algorithm.accept(visitor);
-		this.normalisationParameter = visitor.getResult();
+		this.normalisationParameter = (Double) visitor.getResult();
 
 		return this.normalisationParameter;
 	}
